Add prototype employee recommendation algorithm with datasets
This Prototype Recommendation algorithm is based on item-based
collaborative filtering algorithm. Information about businesses from
the Yelp (2022) dataset has been used as the dataset for this
algorithm.

URL to Yelp (2022) dataset:
https://www.kaggle.com/datasets/yelp-dataset/yelp-dataset

This prototype recommendation algorithm has been built with the
intention of understanding the workflow of the recommendation algorithm
for the given usecase of recommending best employees for a selected
service category. Therefore, this implementation is not efficient in
many aspects such as resource consumption.

This algorithm implementation has been adapted from the implementations
provided in a Linkedin learning course accessible via the following URL:
https://www.linkedin.com/learning/building-recommender-systems-with-machine-learning-and-ai

Five large dataset files (.json and .csv) have been ignored from tracking within
this repository. Please refer to the "Datasets" section at the end of .gitignore
file for more.