Fix the bug of incorrectly parsing dataset files by the ML model
Datasets/ratings_with_splitted_bCategories.csv and Datasets/business.csv
are the two dataset files currently in use for the ML model.

In an earlier commit, all the non-numeric values in the .csv dataset files
were explicitly converted as strings or quoted to avoid the bug of creating
invalid values (#NAME?) in the "business_id" column during data preprocessing
activities. That commit has led to parsing the content in .csv dataset files
incorrectly by the ML model.

With this commit, that issue has been resolved. Now the explicitly quoting process
is applied only on the non-numeric values in .csv files that are used as INPUT for
data preprocessing activites (e.g. yelp_academic_dataset_business.csv, ratings.csv).

Following are the reduced versions of the above dataset files:
    Datasets/Reduced_ratings_with_splitted_bCategories.csv
    Datasets/Reduced_business.csv